[FIX] mail: remove constraint email_from from template
Create an Email Template without email_from
Create Server Action with Send Email
Establish email template as one created

User will be blocked because the template does not have a valid
'email_from'. This does not seems to be useful since the send action is
run via contextual menu, so the message by default will have the address
of the user performing the action.

opw-2213526

closes odoo/odoo#47703

X-original-commit: e083278fe18d491c487e6af4c9ac923acd36c74a
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>