[FIX] website: use automatic color for transparent navbars
When enabling a transparent navbar on a specific page, the user can then
choose a bg color (class) to apply on the transparent navbar. That class
comes with an automatic text color... which was not correctly working
because BS4 overrides colors for nav-links.

Part of https://github.com/odoo/odoo/pull/45351
task-2151408

closes odoo/odoo#47896

X-original-commit: 0397838fbdee8e3eb2240d01300ef2f8263a7c02
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>