[FIX] sale_coupon: validity duration of coupons
- Create a Promotion Program with
  'Applicability' set to 'Apply On Next Order'
  'Promo Code Usage' set to 'Use a code', set a code (not required,
  only for easiness)
- Create a SO, use the code and validate
- In the Promotion Programs, open the coupon list

The Expiration Date is set to tomorrow.

The Expiration Date is not useful in this case, and it is computed only
because the default value of `validity_duration` is 1. However, the user
is not able to change it since `maximum_use_number` is used for
programs instead.

We set a default value to zero to avoid any confusing computation.

opw-2210794

closes odoo/odoo#47779

X-original-commit: 44e71f9192d8711cbaeac0088db55c3ff26813a1
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>