[IMP] crm, mail: clarify the alias configuration in crm.team view
This commit rename the button that redirect to the configuration panel
whare the alias configuration can be done (External email server).

The name was unclear as this is not a domain name that we can configure but
an alias instead. Behind the alias, there is a custom domain, though.

Task ID: 2214653
PR #48023

Related: odoo/enterprise#9367
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>