[FIX] sale: correct dependency for SO line currency
Ensure that the currency of a line always matches the currency of its
order; indeed, with the previous depends, changing the currency of an
order (through its pricelist) would not trigger the recompute of the SO
line's currency.

Was initially fixed in 2dfc0ea8c then reverted at 34e679598 for breaking
a test. I'd rather have the fix and adapt the test though, since this
breaks excpected relations.

Closes #48156

closes odoo/odoo#49491

X-original-commit: 2c8bb518ca6dde0e8e2a5c13531425afff1e22f3
Related: odoo/enterprise#9844
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>